cmd/keeper, cmd/utils, metrics: add zisk target using tamago
