+++ cmd/keeper/dummy_fastcache/go.mod
@@ -1,0 +1,3 @@
+module github.com/ethereum/go-ethereum/cmd/keeper/dummy_fastcache
+
+go 1.24.0
