+++ cmd/keeper/stubs.go
@@ -14,8 +14,8 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with the go-ethereum library. If not, see <http://www.gnu.org/licenses/>.
 
-//go:build !example && !ziren && !wasm && !womir
-// +build !example,!ziren,!wasm,!womir
+//go:build !example && !ziren && !wasm && !womiri && !zisk
+// +build !example,!ziren,!wasm,!womiri,!zisk
 
 package main
