+++ metrics/cpu_enabled.go
@@ -14,8 +14,8 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with the go-ethereum library. If not, see <http://www.gnu.org/licenses/>.
 
-//go:build !ios && !js && !wasip1 && !tinygo
-// +build !ios,!js,!wasip1,!tinygo
+//go:build !ios && !js && !wasip1 && !tinygo && !tamago
+// +build !ios,!js,!wasip1,!tinygo,!tamago
 
 package metrics
