+++ metrics/metrics.go
@@ -9,7 +9,6 @@ package metrics
 import (
 	"runtime/metrics"
 	"runtime/pprof"
-	"time"
 )
 
 var (
@@ -101,104 +100,5 @@ func readRuntimeStats(v *runtimeStats) {
 		case "/sched/latencies:seconds":
 			v.SchedLatency = s.Value.Float64Histogram()
 		}
-	}
-}
-
-// CollectProcessMetrics periodically collects various metrics about the running process.
-func CollectProcessMetrics(refresh time.Duration) {
-	// Short circuit if the metrics system is disabled
-	if !metricsEnabled {
-		return
-	}
-
-	// Create the various data collectors
-	var (
-		cpustats  = make([]CPUStats, 2)
-		diskstats = make([]DiskStats, 2)
-		rstats    = make([]runtimeStats, 2)
-	)
-
-	// This scale factor is used for the runtime's time metrics. It's useful to convert to
-	// ns here because the runtime gives times in float seconds, but runtimeHistogram can
-	// only provide integers for the minimum and maximum values.
-	const secondsToNs = float64(time.Second)
-
-	// Define the various metrics to collect
-	var (
-		cpuSysLoad            = GetOrRegisterGauge("system/cpu/sysload", DefaultRegistry)
-		cpuSysWait            = GetOrRegisterGauge("system/cpu/syswait", DefaultRegistry)
-		cpuProcLoad           = GetOrRegisterGauge("system/cpu/procload", DefaultRegistry)
-		cpuSysLoadTotal       = GetOrRegisterCounterFloat64("system/cpu/sysload/total", DefaultRegistry)
-		cpuSysWaitTotal       = GetOrRegisterCounterFloat64("system/cpu/syswait/total", DefaultRegistry)
-		cpuProcLoadTotal      = GetOrRegisterCounterFloat64("system/cpu/procload/total", DefaultRegistry)
-		cpuThreads            = GetOrRegisterGauge("system/cpu/threads", DefaultRegistry)
-		cpuGoroutines         = GetOrRegisterGauge("system/cpu/goroutines", DefaultRegistry)
-		cpuSchedLatency       = getOrRegisterRuntimeHistogram("system/cpu/schedlatency", secondsToNs, nil)
-		memPauses             = getOrRegisterRuntimeHistogram("system/memory/pauses", secondsToNs, nil)
-		memAllocs             = GetOrRegisterMeter("system/memory/allocs", DefaultRegistry)
-		memFrees              = GetOrRegisterMeter("system/memory/frees", DefaultRegistry)
-		memTotal              = GetOrRegisterGauge("system/memory/held", DefaultRegistry)
-		heapUsed              = GetOrRegisterGauge("system/memory/used", DefaultRegistry)
-		heapObjects           = GetOrRegisterGauge("system/memory/objects", DefaultRegistry)
-		diskReads             = GetOrRegisterMeter("system/disk/readcount", DefaultRegistry)
-		diskReadBytes         = GetOrRegisterMeter("system/disk/readdata", DefaultRegistry)
-		diskReadBytesCounter  = GetOrRegisterCounter("system/disk/readbytes", DefaultRegistry)
-		diskWrites            = GetOrRegisterMeter("system/disk/writecount", DefaultRegistry)
-		diskWriteBytes        = GetOrRegisterMeter("system/disk/writedata", DefaultRegistry)
-		diskWriteBytesCounter = GetOrRegisterCounter("system/disk/writebytes", DefaultRegistry)
-	)
-
-	var lastCollectTime time.Time
-
-	// Iterate loading the different stats and updating the meters.
-	now, prev := 0, 1
-	for ; ; now, prev = prev, now {
-		// Gather CPU times.
-		ReadCPUStats(&cpustats[now])
-		collectTime := time.Now()
-		secondsSinceLastCollect := collectTime.Sub(lastCollectTime).Seconds()
-		lastCollectTime = collectTime
-		if secondsSinceLastCollect > 0 {
-			sysLoad := cpustats[now].GlobalTime - cpustats[prev].GlobalTime
-			sysWait := cpustats[now].GlobalWait - cpustats[prev].GlobalWait
-			procLoad := cpustats[now].LocalTime - cpustats[prev].LocalTime
-			// Convert to integer percentage.
-			cpuSysLoad.Update(int64(sysLoad / secondsSinceLastCollect * 100))
-			cpuSysWait.Update(int64(sysWait / secondsSinceLastCollect * 100))
-			cpuProcLoad.Update(int64(procLoad / secondsSinceLastCollect * 100))
-			// increment counters (ms)
-			cpuSysLoadTotal.Inc(sysLoad)
-			cpuSysWaitTotal.Inc(sysWait)
-			cpuProcLoadTotal.Inc(procLoad)
-		}
-
-		// Threads
-		cpuThreads.Update(int64(threadCreateProfile.Count()))
-
-		// Go runtime metrics
-		readRuntimeStats(&rstats[now])
-
-		cpuGoroutines.Update(int64(rstats[now].Goroutines))
-		cpuSchedLatency.update(rstats[now].SchedLatency)
-		memPauses.update(rstats[now].GCPauses)
-
-		memAllocs.Mark(int64(rstats[now].GCAllocBytes - rstats[prev].GCAllocBytes))
-		memFrees.Mark(int64(rstats[now].GCFreedBytes - rstats[prev].GCFreedBytes))
-
-		memTotal.Update(int64(rstats[now].MemTotal))
-		heapUsed.Update(int64(rstats[now].MemTotal - rstats[now].HeapUnused - rstats[now].HeapFree - rstats[now].HeapReleased))
-		heapObjects.Update(int64(rstats[now].HeapObjects))
-
-		// Disk
-		if ReadDiskStats(&diskstats[now]) == nil {
-			diskReads.Mark(diskstats[now].ReadCount - diskstats[prev].ReadCount)
-			diskReadBytes.Mark(diskstats[now].ReadBytes - diskstats[prev].ReadBytes)
-			diskWrites.Mark(diskstats[now].WriteCount - diskstats[prev].WriteCount)
-			diskWriteBytes.Mark(diskstats[now].WriteBytes - diskstats[prev].WriteBytes)
-			diskReadBytesCounter.Inc(diskstats[now].ReadBytes - diskstats[prev].ReadBytes)
-			diskWriteBytesCounter.Inc(diskstats[now].WriteBytes - diskstats[prev].WriteBytes)
-		}
-
-		time.Sleep(refresh)
 	}
 }
