+++ metrics/metrics_tamago.go
@@ -1,0 +1,12 @@
+//go:build tamago
+// +build tamago
+
+package metrics
+
+import (
+	"time"
+)
+
+// CollectProcessMetrics periodically collects various metrics about the running process.
+func CollectProcessMetrics(time.Duration) {
+}
